fix: [minor] Avoid unnecessary typecasting.
We don't need to typecast identifiers from
their base to type to same type again. This
is not a bug and compiler is fine to skip
it but it is better to avoid if not needed.

+++ cmd/xl-v1-utils.go
@@ -117,7 +117,7 @@ func hashOrder(key string, cardinality int) []int {
 	nums := make([]int, cardinality)
 	keyCrc := crc32.Checksum([]byte(key), crc32.IEEETable)
 
-	start := int(uint32(keyCrc)%uint32(cardinality)) | 1
+	start := int(keyCrc%uint32(cardinality)) | 1
 	for i := 1; i <= cardinality; i++ {
 		nums[i-1] = 1 + ((start + i) % cardinality)
 	}
